+++ src/function/gds/weakly_connected_components.cpp
@@ -18,7 +18,7 @@ using namespace kuzu::graph;
 namespace kuzu {
 namespace function {
 
-class WeaklyConnectedComponentLocalState : public GDSLocalState {
+class WeaklyConnectedComponentLocalState {
 public:
     explicit WeaklyConnectedComponentLocalState(main::ClientContext* context) {
         auto mm = context->getMemoryManager();
@@ -90,7 +90,6 @@ public:
     }
 
     void exec(processor::ExecutionContext*) override {
-        auto wccLocalState = localState->ptrCast<WeaklyConnectedComponentLocalState>();
         auto graph = sharedState->graph.get();
         visitedMap.clear();
         auto groupID = 0;
@@ -105,7 +104,7 @@ public:
                 findConnectedComponent(nodeID, groupID++, *scanState);
             }
         }
-        wccLocalState->materialize(graph, visitedMap, *sharedState->fTable);
+        localState->materialize(graph, visitedMap, *sharedState->fTable);
     }
 
     std::unique_ptr<GDSAlgorithm> copy() const override {
@@ -128,6 +127,8 @@ private:
 
 private:
     common::node_id_map_t<int64_t> visitedMap;
+    // TODO(Semih): Rename to WCCOutputs.
+    std::unique_ptr<WeaklyConnectedComponentLocalState> localState;
 };
 
 function_set WeaklyConnectedComponentsFunction::getFunctionSet() {
