+++ src/include/function/gds/gds_function_collection.h
@@ -10,6 +10,24 @@ struct WeaklyConnectedComponentsFunction {
 
     static function_set getFunctionSet();
 };
+struct AllSPDestinationsFunction {
+    static constexpr const char* name = "ALL_SP_DESTINATIONS";
+
+    static function_set getFunctionSet();
+};
+
+
+struct AllSPLengthsFunction {
+    static constexpr const char* name = "ALL_SP_LENGTHS";
+
+    static function_set getFunctionSet();
+};
+
+struct AllSPPathsFunction {
+    static constexpr const char* name = "ALL_SP_PATHS";
+
+    static function_set getFunctionSet();
+};
 
 struct SingleSPDestinationsFunction {
     static constexpr const char* name = "SINGLE_SP_DESTINATIONS";
